Proposal for modifying CanRetainChecker to apply exclusively to RetainedStateRegistry (#1934)
This work is a proposed change based on the discussion in the #1794 PR
regarding whether `CanRetainChecker` should apply exclusively to the
`RetainedStateRegistry` (see [this
discussion](https://github.com/slackhq/circuit/pull/1794#discussion_r1943388695)).

Based on my review of #834, I believe that `CanRetainChecker` was
originally intended solely for the `RetainedStateRegistry`. However,
once it started being provided as a composition local via
`LocalCanRetainChecker`, we had to handle its partial application to
`rememberRetained` in certain scopes. This has gradually complicated the
logic related to state retention.

In my view, if partial configuration of `CanRetainChecker` for
`rememberRetained` is truly needed, then rather than providing a
`LocalCanRetainChecker` for a specific scope, it could be addressed by
providing a child `RetainedStateRegistry` for that scope.

Accordingly, I have made the following major changes in this work:

### Removed `LocalCanRetainChecker` and configured `CanRetainChecker`
directly on `RetainedStateRegistry`

As described above, since `CanRetainChecker` should affect only direct
`RetainedStateRegistry` children, I have added `CanRetainChecker` as an
argument to `RetainedStateRegistry` and removed the composition local
that provided `LocalCanRetainChecker`.

Furthermore, because `rememberCanRetainChecker` essentially exists only
for the `continuityRetainedRegistry`, I have renamed it to
`rememberContinuityCanRetainChecker` and relocated its definition.

### Removed `CanRetainChecker` logic from `rememberRetained`

Since `CanRetainChecker` now affects only the registry, I have removed
the CanRetainChecker-related logic from `rememberRetained`.
Additionally, the test cases for switching `CanRetainChecker` that Alex
previously added are no longer valid and have been removed.

### Changes related to `RetainedStateHolder`

Since `RetainedStateHolder` is itself a RetainedStateRegistry, for
consistency I have added CanRetainChecker as an argument. Moreover,
because it is no longer possible to wrap
`RetainedStateHolder.RetainedStateProvider` with a
`LocalCanRetainChecker` implementation, I modified it to call
`removeState` in the same manner as was handled for
`SaveableStateHolder` in #1922.

---------------------------------------------

This is a rather drastic change, so I’m not sure if it will be merged. 
However, I wanted to demonstrate how much the overall state retention
logic can be simplified by removing `LocalCanRetainChecker`.

I’d appreciate your thoughts on this!