Screen based animation overrides (#1901)
## About
Building out a surface where navigation transitions can be overridden at
navigation time based on the type of navigation event and the
source/target screens.

Due to how shared element transitions work with `AnimatedContent`, we
need some ability to completely replace the the transition spec before
the target shared elements are loaded into the composition. Can look at
`PetDetailAnimatedScreenTransform` for an example of this.

## `AnimatedScreenTransform` usage
- This interface allows you to customize the `EnterTransition`, z-index,
and `SizeTransform` for a `Screen` as it becomes the top screen or is
replaced by another `Screen` in the navigation stack.
- You can also customize the `ExitTransition` for a screen that is no
longer the top screen.
- Each can be customized based on the `AnimatedNavEvent` that triggered
the transition.

#### Transition Selection
1. **Get base transform:** Calls the `transitionSpec` on the decorator
to obtain a `ContentTransform` based on the `AnimatedNavEvent`.
1. **Check for screen-specific overrides:** Checks for any
`AnimatedScreenTransform` overrides defined for the target screen and
the initial screen.
1. **Combine overrides:** It combines the base transform, and
screen-specific overrides into a final `ContentTransform`.
Screen-specific overrides take precedence over the base transform.
 
### Example
```kotlin
object PetDetailAnimatedScreenTransform : AnimatedScreenTransform {

  override fun AnimatedContentTransitionScope<AnimatedNavState>.enterTransition(
    animatedNavEvent: AnimatedNavEvent
  ): EnterTransition? {
    // Going to the detail screen
    if (initialState.screen !is HomeScreen) return null
    if (!targetState.screen.isSharedElementDetailScreen()) return null
    return EnterTransition.None
  }

  override fun AnimatedContentTransitionScope<AnimatedNavState>.exitTransition(
    animatedNavEvent: AnimatedNavEvent
  ): ExitTransition? {
    // Going back to the home screen
    if (targetState.screen !is HomeScreen) return null
    if (!initialState.screen.isSharedElementDetailScreen()) return null
    return ExitTransition.None
  }
}
```

The example checks if the user is navigating to the `PetDetailScreen`
from the `HomeScreen`, with the information needed to tie together a
shared element transition. If that is the case no transitions are used,
otherwise it returns null to fallback to the default nav decorator
transition.

# Demo

Notice how on the forward navigation the grid view is sliding out, on
the pop `AndroidPredictiveBackNavDecorator` already removes the
animation.

|Before|After|
|-|-|
|<video
src="https://github.com/user-attachments/assets/59a5f669-de71-44ef-8e1d-481b80b64524"></video>|<video
src="https://github.com/user-attachments/assets/076d494d-2ab8-4976-965e-6578c8a89472"></video>|