Fix reset root  (#1995)
🤞🏻 this fixes #1285 

Underlying issue was caused by a new `MovableContent` getting created
for a backstack entry that was still in the composition due to
animations. The new movable content is effectively the same as the old
one, same composite keys etc, so any keyed `ReusableContent` would get
reused without being disposed. This was causing the check in
`SaveableStateHolderImpl` to be replayed in new `onRemembered` call on
the same reused instance.

Fix is to fully track the movable content being used, and remove it from
our provider reuse map when it is no longer going to be used. Allowing
the content to be correctly reparented.

See #2001 for a sample app an test for this