Fix codegen with screen objects to always use `is` check (#2307)
The current codegen uses an equality check when the screen is an object,
but this means that the screen isn't smart cast and thus fails
compilation if the Presenter tries to have it injected. See #2306 for
details and repro steps.

When using a screen object, adding an explicit cast fixes the issue (ie.
`FooPresenter(screen = screen as FooScreen)`)

~~Using `is` even when the screen is an object ensures it can be smart
cast and thus resolves the issue.~~ (old approach, see PR thread for
history)

This PR closes #2306