Add new short-hand utility for collectAsRetainedState (#2403)
Adding a new safe shortcut for collecting a flow. This came out of team
discussions where we would like to have a one-liner for this, but also
want to make sure that it is safe from recompositions and memory leaks.

For instance, we have a habit of doing the following which _retains_ the
flow instance and has a risk of a memory leak across configuration or
navigation. Note that the `inputs` are not specified here.

```kotlin
    val result by
      rememberRetained { dependency.myFlow() }
        .collectAsRetainedState(initial = null)
```

As per the `PetListScreen` example, we should specify the `dependency`
as an input for `rememberRetained` to prevent the flow from being
retained when the dependency changes. It is not obvious that we should
do this which leads to people missing this important detail.

This PR adds a variant of `collectAsRetainedState` that requires
providing `inputs` in a way that prevents memory leaks of the flow and
enables collection as a single statement

```kotlin
      val result by collectAsRetainedState(dependency, initial = null) { myFlow() }
```